move camelCase functions to a separate file
